--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "CodeGPT.apiKey": "Ollama"
-}
