+++ bluenaas/domains/nexus.py
@@ -49,7 +49,7 @@ class BaseNexusSimulationResource(BaseModel):
     injectionLocation: str
     recordingLocation: list[str] | str
     brainLocation: Any  # TODO Add better type
-    is_draft: bool | None = None
+    isDraft: bool | None = None
     status: str | None = None  # TODO Add better type
 
     class Config:
@@ -64,7 +64,7 @@ class FullNexusSimulationResource(NexusBaseResource):
     injectionLocation: str
     recordingLocation: list[str] | str
     brainLocation: Any  # TODO Add better type
-    is_draft: bool | None = None
+    isDraft: bool | None = None
     status: SimulationStatus | None = None  # TODO Add better type
 
     class Config:
